2293: - extracted workspace Configuration logic and repository import into a new class IdeWorkspaceConfigurer
