+++ cli/src/main/java/com/devonfw/tools/ide/tool/ide/IdeToolCommandlet.java
@@ -1,19 +1,13 @@
 package com.devonfw.tools.ide.tool.ide;
 
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Set;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.devonfw.tools.ide.common.Tag;
 import com.devonfw.tools.ide.context.IdeContext;
-import com.devonfw.tools.ide.io.FileAccess;
 import com.devonfw.tools.ide.process.ProcessMode;
 import com.devonfw.tools.ide.process.ProcessResult;
-import com.devonfw.tools.ide.step.Step;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolInstallRequest;
 import com.devonfw.tools.ide.tool.ToolInstallation;
@@ -25,9 +19,9 @@ import com.devonfw.tools.ide.tool.vscode.Vscode;
 /**
  * {@link ToolCommandlet} for an IDE (integrated development environment) such as {@link Eclipse}, {@link Vscode}, or {@link Intellij}.
  */
-public abstract class IdeToolCommandlet extends PluginBasedCommandlet {
+public abstract class IdeToolCommandlet extends PluginBasedCommandlet implements IdeFeatures {
 
-  private static final Logger LOG = LoggerFactory.getLogger(IdeToolCommandlet.class);
+  private final IdeWorkspaceConfigurer workspaceConfigurer;
 
   /**
    * The constructor.
@@ -40,6 +34,7 @@ public abstract class IdeToolCommandlet extends PluginBasedCommandlet {
 
     super(context, tool, tags);
     assert (hasIde(tags));
+    this.workspaceConfigurer = new IdeWorkspaceConfigurer(context, tool);
   }
 
   private boolean hasIde(Set<Tag> tags) {
@@ -71,66 +66,22 @@ public abstract class IdeToolCommandlet extends PluginBasedCommandlet {
   }
 
   /**
+   * Configure (initialize or update) the workspace for this IDE using the templates from the settings.
+   */
+  @Override
+  public void configureWorkspace() {
+    this.workspaceConfigurer.configureWorkspace();
+  }
+
+  /**
    * @return the {@link Path} to the IDE-specific metadata folder for the {@link IdeContext#getWorkspaceName() current workspace}, located at
    *     {@code $IDE_HOME/.ide/«ide»/«workspace»}. Unlike {@link IdeContext#getWorkspacePath() the workspace path} (which holds the projects to open), this
    *     folder keeps IDE-specific metadata (e.g. {@code .vmoptions} or {@code *.properties} files) out of the workspace so it stays clean and independent of
    *     the IDE being used.
    */
-  protected Path getIdeMetadataPath() {
-
-    return this.context.getIdeHome().resolve(IdeContext.FOLDER_DOT_IDE).resolve(getName()).resolve(this.context.getWorkspaceName());
-  }
-
-  /**
-   * Configure (initialize or update) the workspace for this IDE using the templates from the settings.
-   */
-  protected void configureWorkspace() {
-
-    FileAccess fileAccess = this.context.getFileAccess();
-    Path workspaceFolder = this.context.getWorkspacePath();
-    if (!fileAccess.isExpectedFolder(workspaceFolder)) {
-      LOG.warn("Current workspace does not exist: {}", workspaceFolder);
-      return; // should actually never happen...
-    }
-    Step step = this.context.newStep("Configuring workspace " + workspaceFolder.getFileName() + " for IDE " + this.tool);
-    step.run(() -> doMergeWorkspaceStep(step, workspaceFolder));
-  }
-
-  private void doMergeWorkspaceStep(Step step, Path workspaceFolder) {
-
-    int errors = 0;
-    errors = mergeWorkspace(this.context.getUserHomeIde(), workspaceFolder, errors);
-    errors = mergeWorkspace(this.context.getSettingsPath(), workspaceFolder, errors);
-    errors = mergeWorkspace(this.context.getConfPath(), workspaceFolder, errors);
-    if (errors == 0) {
-      step.success();
-    } else {
-      step.error("Your workspace configuration failed with {} error(s) - see log above.\n"
-          + "This is either a configuration error in your settings git repository or a bug in IDEasy.\n"
-          + "Please analyze the above errors with your team or IDE-admin and try to fix the problem.", errors);
-      this.context.askToContinue(
-          "In order to prevent you from being blocked, you can start your IDE anyhow but some configuration may not be in sync.");
-    }
-  }
-
-  private int mergeWorkspace(Path configFolder, Path workspaceFolder, int errors) {
-
-    int result = errors;
-    result = mergeWorkspaceSingle(configFolder.resolve(IdeContext.FOLDER_WORKSPACE), workspaceFolder, result);
-    result = mergeWorkspaceSingle(configFolder.resolve(this.tool).resolve(IdeContext.FOLDER_WORKSPACE), workspaceFolder, result);
-    return result;
-  }
-
-  private int mergeWorkspaceSingle(Path templatesFolder, Path workspaceFolder, int errors) {
-
-    Path setupFolder = templatesFolder.resolve(IdeContext.FOLDER_SETUP);
-    Path updateFolder = templatesFolder.resolve(IdeContext.FOLDER_UPDATE);
-    if (!Files.isDirectory(setupFolder) && !Files.isDirectory(updateFolder)) {
-      LOG.trace("Skipping empty or non-existing workspace template folder {}.", templatesFolder);
-      return errors;
-    }
-    LOG.debug("Merging workspace templates from {}...", templatesFolder);
-    return errors + this.context.getWorkspaceMerger().merge(setupFolder, updateFolder, this.context.getVariables(), workspaceFolder);
+  @Override
+  public Path getIdeMetadataPath() {
+    return this.workspaceConfigurer.getIdeMetadataPath();
   }
 
   /**
@@ -138,8 +89,8 @@ public abstract class IdeToolCommandlet extends PluginBasedCommandlet {
    *
    * @param repositoryPath the {@link Path} to the repository directory to import.
    */
+  @Override
   public void importRepository(Path repositoryPath) {
-
-    throw new UnsupportedOperationException("Repository import is not yet implemented for IDE " + this.tool);
+    this.workspaceConfigurer.importRepository(repositoryPath);
   }
 }
